animeQ Object Constructor and prototype
The documentation inside the file labels extensions on the prototype object as private and public. This is merely post suggestive documentation, as there is no actual private implementation of methods or properties. This code is merely an example of how complex animations can be achieved with vanilla JavaScript. The actual structure of the object could be redesigned, but I don't see the use for the situations this will be used in